+++ nautilus/auth/forms/__init__.py
@@ -1,0 +1,4 @@
+from .change import ChangePasswordForm
+from .forgot import ForgotPasswordForm
+from .login import LoginForm
+from .register import RegistrationForm
